Create a separate file for user data
